(READ MY COMMENTS) Playerlist framework, some cleanup
I removed some warnings because they can be ignored and annoy me when compiling - they are not likely to cause crashes, considering the code isn't even ra n
I made the AntiCat thing only work if _DEBUG is not defined (VS Debug build), so it will still work in release where ServerCmdKeyValues does not crash you

The "tickbase fix" in hitscan aimbot actually causes more prediction errors than if not used, as such i have commented it out.

You can use the new playerlist by including the header file - I have done everything in there for some reason.

I removed a link to the d3dx9 lib as it only applies to 64 bit, and is therefore not needed. (It also made me unable to compile). - I THINK THIS NEEDS TESTING AS I DONT KNOW IF MY VS IS JUST NOT RIGHT OR IF IT REALLY IS UNNECESSARY